setup/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
redis==5.0.1
pika==1.3.2
PyJWT==2.9.0
python-dotenv==1.0.1

# below is used to insert data into PostgreSQL for testing
psycopg2-binary==2.9.9
